move context to its own dir
