+++ __tests__/product/[id].tsx
@@ -1,4 +1,4 @@
-import { CartContext, CartProvider } from '@/util/CartContext';
+import { CartContext, CartProvider } from '@/context/CartContext';
 import '@testing-library/jest-dom';
 import { fireEvent, render, screen } from '@testing-library/react';
 import Cart from '../../components/Cart/Cart';
